Color codes shared clippings in the sidebar. Alters the recieving user w...

+++ db-init.php
@@ -5,8 +5,10 @@ require 'config.php';
 $sql = mysqli_connect($_DB_HOST_, $_DB_USER_, $_DB_PASS_) or die("A MySQL error has occurred.<br />Error: (" . mysqli_errno($sql) . ") " . mysqli_error($sql));
 
 // Drop the database.
-$query = 'DROP DATABASE ' . $_DB_NAME_;
-mysqli_query($sql, $query) or die("A MySQL error has occurred.<br />Error: (" . mysqli_errno($sql) . ") " . mysqli_error($sql));
+if ($_GET['drop'] == 'true') {
+  $query = 'DROP DATABASE ' . $_DB_NAME_;
+  mysqli_query($sql, $query) or die("A MySQL error has occurred.<br />Error: (" . mysqli_errno($sql) . ") " . mysqli_error($sql));
+}
 
 // Create the database and select it.
 $query = 'CREATE DATABASE IF NOT EXISTS ' . $_DB_NAME_;
@@ -20,6 +22,7 @@ $query = 'CREATE TABLE IF NOT EXISTS ' . $_DB_USERS_TABLE_ . '(
   PASSWORD varchar(255) NOT NULL,
   FNAME varchar(255) NOT NULL,
   LNAME varchar(255) NOT NULL,
+  NOTIFICATION varchar(255),
   PRIMARY KEY (ID)
   )';
 mysqli_query($sql, $query) or die("A MySQL error has occurred.<br />Error: (" . mysqli_errno($sql) . ") " . mysqli_error($sql));
@@ -46,6 +49,16 @@ $query = 'CREATE TABLE IF NOT EXISTS ' . $_DB_CLIPPINGS_TABLE_ . '(
   CONTENT longtext NOT NULL,
   NAME varchar(255),
   SUBTITLE varchar(255),
+  PRIMARY KEY (ID)
+  )';
+mysqli_query($sql, $query) or die("A MySQL error has occurred.<br />Error: (" . mysqli_errno($sql) . ") " . mysqli_error($sql));
+
+// Create the shared clippings table.
+$query = 'CREATE TABLE IF NOT EXISTS ' . $_DB_SHARED_CLIPPINGS_TABLE_ . '(
+  ID int(11) AUTO_INCREMENT,
+  CID int(11) NOT NULL,
+  ORIGCID int(11) NOT NULL,
+  UID int(11) NOT NULL,
   PRIMARY KEY (ID)
   )';
 mysqli_query($sql, $query) or die("A MySQL error has occurred.<br />Error: (" . mysqli_errno($sql) . ") " . mysqli_error($sql));
